API call in HomePage
